Update Dockerfile to use local copy of IVCAP Python SDK
The Dockerfile has been split into a multi-stage build.

The first stage installs a build environment and copies the IVCAP Python SDK
source files into the container. By referencing a local copy of the IVAP Python
SDK, no SSH keys need to be copied into the container. A distribution wheel for
the IVCAP Python SDK is built using the copied source code.

The second stage copies the distribution wheel and installs the IVCAP Python
Service SDK and the example requirements.