Added make target 'docker-publish-local'.
Added a make target for publishing a 'locally' built Docker image. The
'docker-publish' target automatically calls 'docker-build'. The target
'docker-build' relies on SSH keys within the build context.

Use this pattern in 'docker-publish-local' to call 'docker-build-local'.

All the '-local' targets have been moved to the end of the Makefile so
that they can be maintained or deleted in a block.