feat(example): Package and deploy CLI tool
This commit introduces a new example (found in primitive)
showing how to package up primitive, a command line application written
in Go, and make it available via an IVCAP service.

I put this together when I first joined the team and was learning how
to use IVCAP. This seems the best place to store this example.